Move some namespaces around
