+++ EliteAPI/Models/Entities/Accounts/EliteAccount.cs
@@ -1,10 +1,12 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using EliteAPI.Models.Entities.Events;
 using EliteAPI.Models.Entities.Hypixel;
 
-namespace EliteAPI.Models.Entities;
+namespace EliteAPI.Models.Entities.Accounts; 
 
-public class AccountEntity
+[Table("Accounts")]
+public class EliteAccount
 {
     [Key]
     public required ulong Id { get; set; }
@@ -27,6 +29,7 @@ public class AccountEntity
     [Column(TypeName = "jsonb")]
     public EliteSettings Settings { get; set; } = new();
     
+    public List<EventMember> EventEntries { get; set; } = new();
     public List<MinecraftAccount> MinecraftAccounts { get; set; } = new();
 }
 
@@ -77,48 +80,4 @@ public class EliteSettings
 {
     public string DefaultPlayerUuid { get; set; } = string.Empty;
     public bool HideDiscordTag { get; set; } = false;
-}
-
-public class MinecraftAccount
-{
-    [Key]
-    public required string Id { get; set; }
-    public required string Name { get; set; }
-    public ulong? AccountId { get; set; }
-    
-    public bool Selected { get; set; }
-    
-    public PlayerData? PlayerData { get; set; }
-
-    [Column(TypeName = "jsonb")]
-    public List<MinecraftAccountProperty> Properties { get; set; } = new();
-    
-    public AccountFlag Flags { get; set; } = AccountFlag.None;
-    public bool IsBanned => Flags.HasFlag(AccountFlag.Banned);
-    
-    [Column(TypeName = "jsonb")]
-    public List<FlagReason>? FlagReasons { get; set; }
-
-    public long LastUpdated { get; set; }
-    public long ProfilesLastUpdated { get; set; }
-    public long PlayerDataLastUpdated { get; set; }
-}
-
-public class MinecraftAccountProperty
-{
-    public required string Name { get; set; }
-    public required string Value { get; set; }
-}
-
-public enum AccountFlag : ushort {
-    None = 0,
-    AutoFlag = 1,
-    Banned = 2,
-}
-
-public class FlagReason {
-    public AccountFlag Flag { get; set; }
-    public string Reason { get; set; } = string.Empty;
-    public string Proof { get; set; } = string.Empty;
-    public long Timestamp { get; set; }
 }
